Hello world with Compose
