+++ .idea/artifacts/web_js.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="web-js">
+    <output-path>$PROJECT_DIR$/application/web/build/libs</output-path>
+    <root id="archive" name="web-js.jar">
+      <element id="module-output" name="Writeopia.application.web.jsMain" />
+    </root>
+  </artifact>
+</component>
