+++ .idea/artifacts/web_wasm.xml
@@ -1,0 +1,6 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="web-wasm">
+    <output-path>$PROJECT_DIR$/application/web/build/libs</output-path>
+    <root id="archive" name="web-wasm.jar" />
+  </artifact>
+</component>
