+++ application/web/build.gradle.kts
@@ -1,0 +1,30 @@
+plugins {
+    kotlin("multiplatform")
+    id("org.jetbrains.compose")
+
+}
+
+kotlin {
+    js(IR) {
+        browser()
+        binaries.executable()
+    }
+    sourceSets {
+        val jsMain by getting {
+            dependencies {
+                api(compose.runtime)
+                api(compose.foundation)
+                api(compose.ui)
+                api(compose.material)
+                api(compose.material3)
+                api(compose.materialIconsExtended)
+                @OptIn(org.jetbrains.compose.ExperimentalComposeLibrary::class)
+                api(compose.components.resources)
+            }
+        }
+    }
+}
+
+compose.experimental {
+    web.application {}
+}
