+++ application/web/src/jsMain/kotlin/Main.kt
@@ -1,0 +1,17 @@
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.window.Window
+import org.jetbrains.skiko.wasm.onWasmReady
+
+fun main() {
+    onWasmReady {
+        Window("Compose Rich Editor") {
+            Box(Modifier.fillMaxSize()) {
+                Text("putz putz putz")
+            }
+        }
+    }
+}
+
