+++ application/web/src/jsMain/resources/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Writeopia web</title>
+    <style>
+        html, body {
+            margin: 0;
+            padding: 0;
+            overflow: hidden;
+        }
+    </style>
+</head>
+<body>
+<canvas id="ComposeTarget"></canvas>
+<script>
+    var canvas = document.getElementById("ComposeTarget");
+    canvas.width = window.innerWidth;
+    canvas.height = window.innerHeight;
+</script>
+<script src="skiko.js"></script>
+<script src="web.js"></script>
+</body>
+</html>
