+++ application/web/src/wasmMain/resources/index.html
