--- application/webApp/src/jsMain/resources/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Writeopia WebApp</title>
-    <script src="skiko.js"></script>
-</head>
-<body>
-<canvas id="ComposeTarget"></canvas>
-<script src="webApp.js"> </script>
-</body>
-</html>
