+++ settings.gradle.kts
@@ -37,4 +37,4 @@ include(":application:features:auth")
 include(":application:persistence")
 include(":application:desktopApp")
 include(":backend:intronotes")
-include(":application:webApp")
+include(":application:web")
